+++ app/Relationships/AlertRelationships.php
@@ -2,64 +2,13 @@
 
 namespace App\Relationships;
 
-use App\Comment;
-use App\Photo;
-use App\User;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\BelongsToMany;
-use Illuminate\Database\Eloquent\Relations\MorphMany;
-
 trait AlertRelationships
 {
     /**
-     * Get all of the owning alert models.
+     * The related alert resource.
      */
     public function alert()
     {
         return $this->morphTo();
-    }
-
-    /**
-     * The related User.
-     *
-     * @return BelongsTo
-     */
-    public function user()
-    {
-        return $this->belongsTo(User::class, 'user_uuid', 'uuid');
-    }
-
-    /**
-     * The related Comments.
-     *
-     * @return MorphMany
-     */
-    public function comments()
-    {
-        return $this->morphMany(Comment::class, 'commentable',
-            'commentable_type',
-            'commentable_id',
-            'uuid'
-        );
-    }
-
-    /**
-     * The related Photo.
-     *
-     * @return BelongsTo
-     */
-    public function photo()
-    {
-        return $this->belongsTo(Photo::class, 'photo_uuid', 'uuid');
-    }
-
-    /**
-     * The related Photos.
-     *
-     * @return BelongsToMany
-     */
-    public function photos()
-    {
-        return $this->morphToMany(Photo::class, 'photoable','photoables','photo_uuid','photoable_id','uuid','uuid');
     }
 }
