+++ app/Relationships/Commons/BelongsToUser.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Relationships\Commons;
+
+use App\User;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+
+trait BelongsToUser
+{
+    /**
+     * The resource related user owner.
+     *
+     * @return BelongsTo
+     */
+    public function user()
+    {
+        return $this->belongsTo(User::class, 'user_uuid', 'uuid');
+    }
+}
