+++ app/Relationships/Commons/Groupable.php
@@ -1,0 +1,20 @@
+<?php
+
+
+namespace App\Relationships\Commons;
+
+use App\Group;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+
+trait Groupable
+{
+    /**
+     * The resource related groups.
+     *
+     * @return BelongsToMany
+     */
+    public function groups()
+    {
+        return $this->morphToMany(Group::class,'groupable','groupables','groupable_id','group_uuid','uuid','uuid')->withPivot(['is_admin','status','sender_uuid',])->withTimestamps();
+    }
+}
