+++ app/Relationships/GeofenceRelationships.php
@@ -4,18 +4,15 @@ namespace App\Relationships;
 
 use App\Access;
 use App\Device;
-use App\Group;
 use App\Pet;
-use App\Photo;
 use App\User;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 
 trait GeofenceRelationships
 {
     /**
-     * The related Geofenceable.
+     * The related geofenceable resource.
      * @codeCoverageIgnore
      */
     public function geofenceable()
@@ -23,9 +20,8 @@ trait GeofenceRelationships
         return $this->morphTo('geofenceable','geofenceable_type','geofenceable_id','geofence_uuid');
     }
 
-
     /**
-     * The related Devices.
+     * The related geofence devices.
      *
      * @return BelongsToMany
      */
@@ -35,7 +31,7 @@ trait GeofenceRelationships
     }
 
     /**
-     * The related Pets.
+     * The related geofence pets.
      *
      * @return BelongsToMany
      */
@@ -45,7 +41,7 @@ trait GeofenceRelationships
     }
 
     /**
-     * The related Users.
+     * The related geofence users.
      *
      * @return BelongsToMany
      */
@@ -55,51 +51,12 @@ trait GeofenceRelationships
     }
 
     /**
-     * The related User.
+     * The related geofence accesses.
      *
-     * @return BelongsTo
-     */
-    public function user()
-    {
-        return $this->belongsTo(User::class, 'photo_uuid', 'uuid');
-    }
-
-    /**
-     * The related Photo
-     *
-     * @return BelongsTo
-     */
-    public function photo()
-    {
-        return $this->belongsTo(Photo::class, 'photo_uuid', 'uuid');
-    }
-
-    /**
-     * The related Photos.
-     *
-     * @return BelongsToMany
-     */
-    public function photos()
-    {
-        return $this->morphToMany(Photo::class,'photoable','photoables','photoable_id','photo_uuid','uuid','uuid');
-    }
-
-    /**
      * @return HasMany
      */
     public function accesses()
     {
         return $this->hasMany(Access::class, 'geofence_uuid', 'uuid');
-    }
-
-    /**
-     * The related Groups.
-     *
-     * @return BelongsToMany
-     */
-    public function groups()
-    {
-        return $this->morphToMany(Group::class,'groupable','groupables','groupable_id','group_uuid','uuid','uuid')
-            ->withPivot(['is_admin','status','sender_uuid'])->withTimestamps();
     }
 }
