+++ app/Resources/ActionResource.php
@@ -10,6 +10,8 @@ trait ActionResource
      * Get available alert reactions.
      *
      * @return string
+     * @codeCoverageIgnore TODO Support action reactions
+     * @body Increase code coverage support enabling the action reactions. Just add routes and tests.
      */
     public function getAvailableReactions()
     {
@@ -21,6 +23,8 @@ trait ActionResource
      *
      * @param Request $request
      * @return array
+     * @codeCoverageIgnore
+     * @justify Actions can't be updated, are system generated.
      */
     public function updateRules(Request $request)
     {
@@ -32,6 +36,8 @@ trait ActionResource
      *
      * @param Request $request
      * @return array
+     * @codeCoverageIgnore
+     * @justify Actions can't be created by request, are system generated.
      */
     public function storeRules(Request $request)
     {
