+++ app/Resources/LocationResource.php
@@ -10,6 +10,8 @@ trait LocationResource
      * Get available location reactions.
      *
      * @return string
+     * @codeCoverageIgnore TODO Add location reactions
+     * @body Increase code coverage support enabling the alert reactions. Just add routes and tests.
      */
     public function getAvailableReactions()
     {
@@ -22,6 +24,7 @@ trait LocationResource
      * @param Request $request
      * @return array
      * @codeCoverageIgnore
+     * @justify Locations are device generated, can't be modified by user.
      */
     public function updateRules(Request $request)
     {
@@ -34,6 +37,7 @@ trait LocationResource
      * @param Request $request
      * @return array
      * @codeCoverageIgnore
+     * @justify Locations are device generated, can't be predefined validation rules.
      */
     public function storeRules(Request $request)
     {
